requirements.txt
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
dataclasses==0.8
python-dotenv
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.62.0
h11==0.11.0
idna==2.10
motor==2.3.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pymongo==3.11.2
python-jose==3.2.0
python-multipart==0.0.5
requests
rsa==4.6
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.2
git+git://github.com/sendinblue/APIv3-python-library.git